Update WebUI. Use WebUIAggregatorAPI to get aggregated information about products including stock levels. In ProductItemPartial show discount, original price, price with discount, how many items were sold, if the item is available on stock. Update Order, Cart, Product and CheckOut pages.

+++ Common/Models/Warehouse/StockItem.cs
@@ -22,9 +22,9 @@ namespace Common.Models.Warehouse
         public ushort Quantity { get; set; }
 
         [Required]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
-        public decimal Discount { get; set; }
+        public double Discount { get; set; }
 
         public ushort Sold { get; set; }
